Use "2.8-stable-eos" branch for building NFS ganesha
Some of the NFS ganesha ACL apis/symbols are added to "ver" file
in order to export them. These symbols are used in kvsfs-ganesha
acl code. A new NFS ganesha branch - 2.8-stable-eos - has been
forked out for this. This patch makes sure that this branch will
be used for bootstrapping NFS ganesha.