EOS-4144: Remove redundant pre-condition.
The resize() function should be able to handle
the case where the file is empty even when
the new size is also zero.

This pre-condition blocks the QA team from testing
the workaround proposed for EOS-4144. It was first
proposed in the scope of EOS-4776 but due to being
stuck in design discussions, this fix was buried
along with the EOS-4776.

NOTE: This change does not include EOS-4776 fixes,
therefore any errors (including ENOSPC) are not
propogated to the upper layers, and because of that
any testcase that includes 100% allocated Mero storage
will not show the expected behavior.