EOS-7008: Implementation of kvtree attach, detach, lookup and iter_children
Problem: Kvtree APIs for node attach/detach were required to link a node to
the kvtree and unlinking it. Similarly there was a requirement to do a lookup
for this linked child and iterate over all the children of a parent node.
Solution: Implemented following APIs kvnode_attach(), kvnode_detach(),
kvnode_lookup(), kvnode_iter_children() which returns a kvnode through each
iteration by the callback.