EOS-6837: Set/Get/Modify/Delete ACL on a directory
Problem description:
Currently, eos-fs uses xattrs for storing an ACL for a file. Every xattr (and a file) is
identifued by a unique 128 bit oid. A "directory" does not have a 128 bit oid currently.
They are only identified by 64bit inode. For ACL to work with directories, a directory
 should have an 128 bit oid associated with it.

Fix:
- Added code for mapping a directory inode to 128 bit oid. This happens when directory is created.
- Created a common "library" which contains the definition of these test cases that both file xattr
and dir xattr ut can use.
- Fixed another small bug in dir_ops UT, in dir_test_teardown()